# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Bus Payments & Rewards API
  description: Comprehensive API for bus payment system with offers, rewards, bookings, and refunds
  version: 1.0.0
servers:
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: '5006'
        description: Server port number
...
paths:
...
  /api/merchants/{merchant_id}/history:
    get:
      tags:
        - Merchants
      summary: Get merchant offer history
      parameters:
        - in: path
          name: merchant_id
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per_page
          schema:
            type: integer
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: offer_id
          schema:
            type: integer
      responses:
        '200':
          description: Merchant offer usage history
        '404':
          description: Merchant not found
...
  /api/rewards/customer/{customer_id}/history:
    get:
      tags:
        - Rewards
      summary: Get redemption history
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            default: 10
        - in: query
          name: start_date
          schema:
            type: string
            format: date-time
        - in: query
          name: end_date
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Redemption history
        '400':
          description: Invalid parameters
        '404':
          description: Customer not found
        '500':
          description: Internal server error
...